+++ FROST-Server.SQLjooq/src/main/java/de/fraunhofer/iosb/ilt/frostserver/persistence/pgjooq/JooqAbstractPersistenceManager.java
@@ -241,6 +241,7 @@ public abstract class JooqAbstractPersistenceManager extends AbstractPersistence
         return connectionProvider;
     }
 
+    @Override
     public <R extends Record> ResultQuery<R> setTimeout(ResultQuery<R> query) {
         if (persistenceSettings.isTimeoutQueries()) {
             query.queryTimeout(persistenceSettings.getQueryTimeout());